+++ spec/generators/twirp_generator_spec.rb
@@ -1,15 +1,13 @@
-require 'spec_helper'
+require 'rails_helper'
 require 'generator_spec'
 
 RSpec.describe TwirpGenerator, type: :generator do
-  destination File.expand_path('../tmp/dummy', __dir__)
+  destination File.expand_path('../../tmp/dummy', __dir__)
   arguments %w(sample)
 
   before(:all) do
     prepare_destination
-    FileUtils.cp_r File.expand_path('dummy', __dir__), File.expand_path('../tmp', __dir__)
-    FileUtils.rm_r File.expand_path('../tmp/dummy/lib/twirp', __dir__), force: true
-    FileUtils.mkdir_p File.expand_path('../tmp/dummy/lib/twirp', __dir__)
+    FileUtils.cp_r File.expand_path('../dummy', __dir__), File.expand_path('../../tmp', __dir__)
     run_generator
   end