+++ CHANGELOG.md
@@ -4,6 +4,38 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.4.1 - 2020-03-30
+
+### Added
+
+- smart service detection - you can use `rails g twirp svc` if you `Svc` or 
+`SvcAPI` service described at `company/service/subservice/version/etc/svc_api.proto`
+- Add `# :nocov:` comment to generated modules to avoid decrease coverage
+- Improve generator console output and error handling
+- `add_api_acronym` configuration option
+
+### Fixed
+
+- `protoc` path was cached on start and didn't reload with spring
+
+## 0.4.0 - 2020-03-24
+
+### Breaking changes
+
+- Client and server proto directory splits from `app/protos` to `rpc` and `rpc/clients` (configurable)
+ 
+### Added
+
+- Added gem configuration and generator to create initial configuration file `rails g twirp:init` (comments inside).
+- Proto source dirs and rb destination dirs now configurable.
+- Added separate generator to run protoc on clients proto files `rails g twirp:clients`.
+- Add acronym API to Rails inflector to correct generate `ServiceAPI` handler from `ServiceApi` as protoc twirp plugin.
+- In the development environment gem uses warn instead of raise errors on incorrect generated code or invalid service routes.  
+
+### Fixed
+
+- Fixed incorrect indent in generated modules.
+
 ## 0.3.2 - 2020-03-12
 
 ### Added